% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aalen-Johansen.R, R/MM.r
\name{aalen_johansen}
\alias{aalen_johansen}
\title{Compute the conditional Aalen-Johansen estimator.}
\usage{
aalen_johansen(data = null, na = null)

aalen_johansen(data = null, na = null)
}
\arguments{
\item{data}{A list of trajectory data for each individual.}

\item{na}{A list of Nelson-Aalen estimators}

\item{x}{A numeric value for conditioning.}

\item{a}{A bandwidth. Defaults to Sheather and Jones selection wrt marginal.}

\item{p}{An integer representing the number of states. The absorbing state is last.}

\item{collapse}{Logical, whether to collapse the last state of the model.}

\item{weights}{A vector of observation weights.}
}
\value{
A list containing the Aalen-Johansen estimator, the Nelson-Aalen estimator, and related quantities.
}
\description{
Compute the conditional Aalen-Johansen estimator.

Compute the Aalen-Johansen estimator based on data in a node. Either data
has to be supplied or a list of Nelson-Aalen estimators should be given
}
